Improve identifier detection
Code such as "p.enabled" got mistaken for a partial number, resulting
in extracting the identifier "nabled". This would trigger an assert
later on because there remains a char suitable for identifiers ("e")
in front of "nabled".

Fix by making use of the detected purpose of each char, which is
determined earlier on during whitespace checking.