Fix Python 2-only issues
Because bytes objects are str in Python 2, creating a bytes(80) object
results in a string "80". This breaks the returned Zopfli buffer which
contained the textual representation of the object instead of the buffer
in the object. Fix by using (mutable) bytearray where needed.

Also fix minor print output issues: Progress isn't printed and requires
flushing, and the progress was printed as an integer.